Jpa database

See merge request idatt2106_2022_02/boco-backend!2

+++ src/main/java/no/ntnu/idatt2106/repository/UserRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.idatt2106.model.DAO.UserDAO;
 
-public interface UserRepository extends JpaRepository<UserDAO, Integer> {
+
+public interface UserRepository extends JpaRepository<UserDAO, Long> {
     UserDAO findByEmail(String email);
 }
